requirements.txt
discord
web3==7.2.0
python-dotenv
